+++ backend/src/main/java/mouda/backend/auth/business/AppleAuthService.java
@@ -12,7 +12,9 @@ import mouda.backend.auth.exception.AuthErrorMessage;
 import mouda.backend.auth.exception.AuthException;
 import mouda.backend.auth.implement.AppleOauthManager;
 import mouda.backend.auth.implement.jwt.AccessTokenProvider;
+import mouda.backend.auth.presentation.request.AppleOauthRequest;
 import mouda.backend.auth.presentation.request.AppleUserInfoRequest;
+import mouda.backend.auth.presentation.response.LoginResponse;
 import mouda.backend.member.domain.LoginDetail;
 import mouda.backend.member.domain.Member;
 import mouda.backend.member.domain.OauthType;
@@ -31,16 +33,17 @@ public class AppleAuthService {
 	private final ObjectMapper objectMapper;
 
 	// TODO: 더 이상 사용하지 않는 로직. 로그인 프로세스 정착 후 제거할 것
-	// public LoginResponse oauthLogin(AppleOauthRequest oauthRequest) {
-	// 	TODO: 사용자 전환 로직 실행 이전부터 애플 소셜 회원가입이 진행되어있음. 현재 카카오 사용자가 전환을 시도하여 애플 로그인하면 같은 애플 로그인 사용자가 두 명이 되면서 에러가 터질 것.
-	// if (oauthRequest.memberId() != null) {
-	// 	String accessToken = loginManager.updateOauth(oauthRequest.memberId(), OauthType.APPLE,
-	// 		member.getSocialLoginId());
-	// 	return new LoginResponse(accessToken);
-	// }
-	// LoginProcessResult result = loginManager.processAppleLogin(member);
-	// return new LoginResponse(result.accessToken());
-	// }
+	public LoginResponse oauthLogin(AppleOauthRequest oauthRequest) {
+		// TODO: 사용자 전환 로직 실행 이전부터 애플 소셜 회원가입이 진행되어있음. 현재 카카오 사용자가 전환을 시도하여 애플 로그인하면 같은 애플 로그인 사용자가 두 명이 되면서 에러가 터질 것.
+		// if (oauthRequest.memberId() != null) {
+		// 	String accessToken = loginManager.updateOauth(oauthRequest.memberId(), OauthType.APPLE,
+		// 		member.getSocialLoginId());
+		// 	return new LoginResponse(accessToken);
+		// }
+		// LoginProcessResult result = loginManager.processAppleLogin(member);
+		// return new LoginResponse(result.accessToken());
+		return null;
+	}
 
 	public String login(String idToken) {
 		String socialLoginId = appleOauthManager.getSocialLoginId(idToken);
