+++ backend/src/main/java/mouda/backend/auth/presentation/controller/AuthController.java
@@ -1,6 +1,5 @@
 package mouda.backend.auth.presentation.controller;
 
-import org.springframework.context.annotation.Profile;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -9,12 +8,12 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import lombok.RequiredArgsConstructor;
-import mouda.backend.aop.logging.ExceptRequestLogging;
 import mouda.backend.auth.business.AppleAuthService;
 import mouda.backend.auth.business.CommonAuthService;
 import mouda.backend.auth.business.GoogleAuthService;
 import mouda.backend.auth.business.KakaoAuthService;
 import mouda.backend.auth.presentation.controller.swagger.AuthSwagger;
+import mouda.backend.auth.presentation.request.AppleOauthRequest;
 import mouda.backend.auth.presentation.request.GoogleOauthRequest;
 import mouda.backend.auth.presentation.request.OauthRequest;
 import mouda.backend.auth.presentation.response.LoginResponse;
@@ -33,7 +32,7 @@ public class AuthController implements AuthSwagger {
 	private final CommonAuthService commonAuthService;
 
 	@Override
-	@PostMapping("/kakao/oauth")
+	@PostMapping("/kakao")
 	public ResponseEntity<RestResponse<LoginResponse>> loginKakaoOauth(@RequestBody OauthRequest oauthRequest) {
 		LoginResponse response = kakaoAuthService.oauthLogin(oauthRequest);
 
@@ -41,27 +40,7 @@ public class AuthController implements AuthSwagger {
 	}
 
 	@Override
-	@PostMapping("/login/anna")
-	@ExceptRequestLogging
-	@Profile(value = {"local", "dev"})
-	public ResponseEntity<RestResponse<LoginResponse>> loginBasicOauthAnna() {
-		LoginResponse response = kakaoAuthService.basicLoginAnna();
-
-		return ResponseEntity.ok().body(new RestResponse<>(response));
-	}
-
-	@Override
-	@PostMapping("/login/hogee")
-	@ExceptRequestLogging
-	@Profile(value = {"local", "dev"})
-	public ResponseEntity<RestResponse<LoginResponse>> loginBasicOauthHogee() {
-		LoginResponse response = kakaoAuthService.basicLoginHogee();
-
-		return ResponseEntity.ok().body(new RestResponse<>(response));
-	}
-
-	@Override
-	@PostMapping("/google/oauth")
+	@PostMapping("/google")
 	public ResponseEntity<RestResponse<LoginResponse>> loginGoogleOauth(
 		@RequestBody GoogleOauthRequest googleOauthRequest) {
 		LoginResponse response = googleAuthService.oauthLogin(googleOauthRequest);
@@ -69,13 +48,13 @@ public class AuthController implements AuthSwagger {
 		return ResponseEntity.ok().body(new RestResponse<>(response));
 	}
 
-	// @Override
-	// @PostMapping("/apple/oauth")
-	// public ResponseEntity<RestResponse<LoginResponse>> loginAppleOauth(@RequestBody AppleOauthRequest oauthRequest) {
-	// 	LoginResponse response = appleAuthService.oauthLogin(oauthRequest);
-	//
-	// 	return ResponseEntity.ok().body(new RestResponse<>(response));
-	// }
+	@Override
+	@PostMapping("/apple/oauth")
+	public ResponseEntity<RestResponse<LoginResponse>> loginAppleOauth(@RequestBody AppleOauthRequest oauthRequest) {
+		LoginResponse response = appleAuthService.oauthLogin(oauthRequest);
+
+		return ResponseEntity.ok().body(new RestResponse<>(response));
+	}
 
 	@Override
 	@DeleteMapping
